Refactored signOutUtils
