+++ __test__/NeighbourhoodShell.test.tsx
@@ -8,7 +8,7 @@ import { signOut } from 'aws-amplify/auth';
 import { NextRouter } from 'next/router';
 import { DataProvider } from '@/contexts/DataContext';
 import { useCurrentUser } from '@/src/hooks/usersCustomHooks';
-import { utilSignOut } from '@/utils/signOutUtils';
+import { handleSignOut } from '@/utils/authUtils';
 import AppPage from '@/app/neighbourhood/page';
 
 const mockLikePost = jest.fn();
@@ -198,7 +198,7 @@ describe('Neighbourhood Shell', () => {
   //1.8
   test('utilSinOut should sign out the user, clear localStorage, navigate to home, and show a notification', async () => {
     // Call the utility function with the mocked router
-    await utilSignOut({ router: routerMock as NextRouter });
+    await handleSignOut({ router: routerMock as NextRouter });
 
     // Assertions to ensure all expected actions were called
     expect(signOut).toHaveBeenCalledWith({ global: true });
